requirements.txt
awesome-slugify==1.6.5
py==1.4.30
pytest==2.7.2
regex==2015.7.19
Unidecode==0.4.18
wheel==0.24.0
